requirements.txt
lampe==0.9.0
mlxp==1.0.0
numpyro==0.18.0
scikit-learn==1.6.1
scipy==1.12.0
pyro-ppl==1.9.1
seaborn
ipython
jupyter
matplotlib
tqdm
pot
seaborn
causalchamber
